Reorganize subclassing of tuple
Now the subclass_tuple parameter in the metaclass is removed. Its
semantics is replaced by two base classes ProgrammbleTuple and
ImmutableClass. The utility methods are all moved into a base classs
_ProgrammableTupleBase, which is mixed in into both the public base
classes.

The determinantion of subclassing tuple is not determined by really
testing if the class is a subclass of tuple.