Allow subclassing ordinary classes
Since it is found that simultaneous subclassing both tuple and the base
class is not possible, due to the conflicting layout exception, Now the
utility methods class is renamed to be a mixin class subclassing the
normal object. And subclassing normal classes are now allowed. This is
to facilitate mixing for both this case and other possible use cases as
well.