Reintroduce slots
It is really highly undesirable to make every object carries a __dict__
for only one attribute. Now the slots are re-introduced by separate
handling of tuple subclasses and non tuple subclasses. `__content__` is
now a property method for tuple subclasses so tuple do not need to be
add anything more.