Remove return with argument in the generator
In Python 3.2, generator with return statement of a value will be
considered an error. So now it is fixed by removing the value. This will
not change the semantics of the code at all, since the value set to the
StopIteration exception is not used here anyway. And it expands the
usability of the code.