Improve decoration of the init function
In the old code, the init function is just changed into a disabled
function for the programmable tuple classes. Now it is disable only when
being called on a programmable tuple class. So it can be used for the
initialization of proxy object when being resolved from the class by the
user.